refactoring:plan조회,삭제,디테일controller
